fix virtualId issue
