+++ src/services/orchestration/orchestrationService.js
@@ -1,16 +1,10 @@
 import axios from "axios";
 import { getLocalData } from "../../utils/constants";
 import config from "../../utils/urlConstants.json";
-import { jwtDecode } from "jwt-decode";
+import { getVirtualId } from "../userservice/userService";
 
 const API_BASE_URL_ORCHESTRATION =
   process.env.REACT_APP_LEARNER_AI_ORCHESTRATION_HOST;
-const TOKEN = localStorage.getItem("apiToken");
-let virtualId;
-if (TOKEN) {
-  const tokenDetails = jwtDecode(TOKEN);
-  virtualId = JSON.stringify(tokenDetails?.virtual_id);
-}
 
 const getHeaders = () => {
   const token = getLocalData("apiToken");
@@ -59,7 +53,7 @@ export const addPointer = async (points, milestone) => {
     const response = await axios.post(
       `${API_BASE_URL_ORCHESTRATION}/${config.URLS.ADD_POINTER}`,
       {
-        userId: virtualId,
+        userId: getVirtualId(),
         sessionId: sessionId,
         points: points,
         language: lang,
@@ -84,7 +78,7 @@ export const createLearnerProgress = async (
 
   try {
     const requestBody = {
-      userId: virtualId,
+      userId: getVirtualId(),
       sessionId: sessionId,
       subSessionId: subSessionId,
       milestoneLevel: milestoneLevel,
@@ -117,7 +111,7 @@ export const addLesson = async ({
     const response = await axios.post(
       `${API_BASE_URL_ORCHESTRATION}/${config.URLS.ADD_LESSON}`,
       {
-        userId: virtualId,
+        userId: getVirtualId(),
         sessionId: sessionId,
         milestone: milestone,
         lesson: lesson,
