+++ src/services/userservice/userService.js
@@ -1,5 +1,6 @@
 import axios from "axios";
 import config from "../../utils/urlConstants.json";
+import { jwtDecode } from "jwt-decode";
 
 const API_HOST_VIRTUAL_ID_HOST = process.env.REACT_APP_VIRTUAL_ID_HOST;
 
@@ -13,4 +14,14 @@ export const fetchVirtualId = async (username) => {
     console.error("Error fetching virtual ID:", error);
     throw error;
   }
+};
+
+export const getVirtualId = () => {
+  const TOKEN = localStorage.getItem("apiToken");
+  let virtualId;
+  if (TOKEN) {
+    const tokenDetails = jwtDecode(TOKEN);
+    virtualId = JSON.stringify(tokenDetails?.virtual_id);
+  }
+  return virtualId;
 };
